feat: Define template tag to store timezone.now in the render context
